CareerCraftt-main/backend/requirements.txt
firebase-admin
fastapi
pdfminer.six
scikit-learn
uvicorn
sqlalchemy
spacy
numpy
python-docx
pypdf
google-genai
requests
python-dotenv
